chore(attach-mock): decompose module, fix minor issues
